Limpiando el proyecto
Moviendo las operaciones a las pruebas, y modificando el espacio para ser mas flexible
