+++ Editor/IEspacio.cs
@@ -2,12 +2,6 @@
 {
     public interface IEspacio
     {
-        public bool AgregarElemento(IElemento elemento);
-
-        public bool SacarElemento(IElemento elemento);
-
-        public bool PuedeAgregarElemento(IElemento elemento);
-
         public void AplicarOperacion(IOperacionElementos operacion);
 
         public void AplicarOperacion(IOperacionEspacios operacion);
