+++ Editor/IOperacionElementos.cs
