+++ Editor/IOperacionElementos.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e008156115fc5574aadd1a96911c32d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
