+++ Editor/IOperacionEspacios.cs
