+++ Editor/IOperacionEspacios.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 455c09f298f5ec0429a8f115e263da1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
