+++ Tests/Editor/ISlotPrueba.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 89c5227a2821d474fbc93b299f1a6aa1
+guid: 1c1a8fe68f24a654f9dc90589585ffd7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
