+++ Tests/Editor/InventarioPrueba.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cfb867c36b079d14e9579fc76cfc8ce7
+guid: f224ea8cd535a934db349ddda6ba8ecf
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
