+++ Tests/Editor/Operaciones/CantidadElementosTotales.cs
