+++ Tests/Editor/Operaciones/ElementosIguales.cs
