+++ Tests/Editor/Operaciones/ElementosIguales.cs.meta
