+++ Tests/Editor/Operaciones/SacarElemento.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a48b76e22b2b3324ab076b07062a73ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
