+++ Tests/Editor/Operaciones/TieneEspacio.cs.meta
